[BE] cleanup : 배틀 밸런스 수정

See merge request s08-final/S08P31E105!163

+++ back/src/main/java/com/repomon/rocketdan/domain/repomon/dto/response/RepomonStatusResponseDto.java
@@ -26,7 +26,7 @@ public class RepomonStatusResponseDto {
 
 	// 내부 공식으로 계산한 수치
 	// 최초에 주사위 굴린 수치도 여기에 반영됨
-	private Integer atk;
+	private Float atk;
 	private Float dodge;
 	private Float def;
 	private Float critical;
@@ -42,7 +42,7 @@ public class RepomonStatusResponseDto {
 	private Integer hitPoint;
 
 	// 증가치
-	private Integer increaseAtk;
+	private Float increaseAtk;
 	private Float increaseDodge;
 	private Float increaseDef;
 	private Float increaseCritical;
